src/atlassian_migration_tool/utils/config_loader.py
"""
Configuration loader utility
"""
import os
from pathlib import Path
from typing import Any

import yaml
from dotenv import load_dotenv


def load_config(config_path: str = "config/config.yaml") -> dict[str, Any]:
    """
    Load configuration from YAML file.

    Args:
        config_path: Path to configuration file

    Returns:
        Configuration dictionary
    """
    # Load environment variables from .env file
    load_dotenv()

    config_file = Path(config_path)

    if not config_file.exists():
        raise FileNotFoundError(
            f"Configuration file not found: {config_path}\n"
            "Copy config/config.example.yaml to config/config.yaml"
        )

    with open(config_file) as f:
        config = yaml.safe_load(f)

    # Replace environment variables
    config = _replace_env_vars(config)

    return config


def _replace_env_vars(obj: Any) -> Any:
    """Recursively replace ${VAR} with environment variables."""
    if isinstance(obj, dict):
        return {k: _replace_env_vars(v) for k, v in obj.items()}
    elif isinstance(obj, list):
        return [_replace_env_vars(item) for item in obj]
    elif isinstance(obj, str) and obj.startswith("${") and obj.endswith("}"):
        var_name = obj[2:-1]
        return os.getenv(var_name, obj)
    return obj
