pyproject.toml
[tool.poetry]
name = "atlassian-migration-tool"
version = "0.1.0"
description = "A comprehensive tool for migrating content from Atlassian Confluence and Jira to open-source alternatives (Wiki.js, OpenProject, GitLab)"
authors = ["David Dunnock <dunnoda@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ddunnock/atlassian-migration-tool"
repository = "https://github.com/ddunnock/atlassian-migration-tool.git"
keywords = ["atlassian", "confluence", "jira", "migration", "wikijs", "openproject", "gitlab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

[tool.poetry.dependencies]
python = "^3.10"
# Atlassian API
atlassian-python-api = "^3.41.0"

# HTTP and API clients
requests = "^2.31.0"
urllib3 = "^2.1.0"
python-gitlab = "^4.4.0"

# HTML and Markdown processing
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
markdownify = "^0.12.0"
html2text = "^2024.2.26"
markdown = "^3.5.0"

# Configuration and data handling
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
toml = "^0.10.2"

# CLI framework
click = "^8.1.7"
rich = "^13.7.0"
typer = "^0.9.0"

# Templating
jinja2 = "^3.1.0"

# Date and time handling
python-dateutil = "^2.8.2"
pytz = "^2024.1"

# Logging and monitoring
loguru = "^0.7.2"
colorlog = "^6.8.0"

# File handling
pathlib2 = "^2.3.7"
aiofiles = "^23.2.0"

# Async support
aiohttp = "^3.9.0"
asyncio = "^3.4.3"

# Web GUI (FastAPI)
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sse-starlette = "^2.0.0"
python-multipart = "^0.0.9"

# Data validation and serialization
jsonschema = "^4.20.0"
cattrs = "^23.2.0"

# Database (for tracking migrations)
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"

# Progress bars and UI
tqdm = "^4.66.0"
alive-progress = "^3.1.5"

# Retry and rate limiting
tenacity = "^8.2.0"
ratelimit = "^2.2.1"

# Security
cryptography = "^42.0.0"
keyring = "^24.3.0"

# Testing (optional but included)
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.0"

# Documentation
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"

[tool.poetry.group.dev.dependencies]
# Code quality
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
isort = "^5.13.0"
pylint = "^3.0.0"

# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.0"
pytest-xdist = "^3.5.0"
faker = "^22.0.0"

# Type stubs
types-requests = "^2.31.0"
types-PyYAML = "^6.0.12"
types-python-dateutil = "^2.8.19"

# Development tools
ipython = "^8.20.0"
ipdb = "^0.13.13"
pre-commit = "^3.6.0"

# Documentation
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"
mkdocstrings = "^0.24.0"

[tool.poetry.scripts]
atlassian-migrate = "atlassian_migration_tool.cli:main"
atlassian-migrate-gui = "atlassian_migration_tool.web.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "atlassian.*",
    "gitlab.*",
    "bs4.*",
    "markdownify.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "confluence: Confluence-specific tests",
    "jira: Jira-specific tests",
    "wikijs: Wiki.js-specific tests",
    "openproject: OpenProject-specific tests",
    "gitlab: GitLab-specific tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
